Stop persisting OicSession in HTTP session when only OIDC state is required
In a CloudBees CI HA setup, I recently upgraded to 4.297.vcddb_d8a_e4694 (including #310). The various changes to OicSecurityRealm broke serialization due to usage of an anonymous inner class (extending OicSession) that gets stored in session.

Storing `OicSession` in session is actually only required for a brief amount of time, only between doCommenceLogin and doFinishLogin. Once the user is logged in, it is no longer necessary to store the whole object. The only thing that needs to be persisted is the state, as it gets used later for logging out.