rework configuration of the plugin
The server configuration (token URLs, scopes etc) are now in separate
describable.

This makes the UX cleaner, and the code cleaner in the realm.

The config.xml is backwards compatable with previous versions, but the casc
format is not.

when in discovery mode via well known endpoint the override scopes are
now used explicitly rather than using the overlapping scopes from the
overrides and those provided by the server