Implement a custom configuration so we can modify HttpRequests
The OidcAuthenticator was not using the resource retreiver to talk to
servers. As such when used against a server with a self signed
certificate and disableTLS checks was set it would still fail.

Whilst we could implement our own Authenticator, there may be other
places where we need to modify the HttpRequest.  Therefore  we just
create a custom configuration that will set the proxy and TLS options as
required.