workaround token refresh nonce issue
Pac4j setups the token validators, which during the refresh lifecycle
will attempt to check an ID tokens nonce.
However a provider should not set the nonce in the idtoken during a
refresh, and in this case the validator fails because the nonce is
missing from the token!

we disable the nonce check for the refresh call.  it can be optionally
re-enabled by setting the system property
org.jenkinsci.plugins.oic.OicSecurityRealm.checkNonceInRefreshFlow to
true.
this is not exposed as a config option in the UI as
1) providers should not be sending the nonce anyway
2) this should be a short lived workaround whilst the issue with the
   library is filed and fixed.