Resuse the original user principal to avoid crumb issues.
as it has been observed that the case of a user may change during a
refreesh flow even though they are the same user, the crumb uses the
Authentications name (principal), which would be different as we use the
returned value.
Rather than using the new value, after checking it is the same id
(according to the ID Strategy) we use the original so that the crumb can
be matched.

maybe fixes: #411