Introduce `OicProperty` extension point
In the context of CloudBees CI, I needed the ability to implement new
optional behaviour customizing the OIDC configuration and client.

The new extension point, named `OicProperty`, allows to capture a state
through `OicPropertyExecution` that can be used for customize both the
OIDC configuration and client.

I have also taken the opportunity to fold some advanced options under
their own class. Ultimately, all advanced fields could be implemeneted
as `OicProperty` but for some of them it can be challenging due to the
way they are currently integrated in the plugin.

Migrated fields
* Allowed token expiration clock skew
* Disable nonce
* Disable token verification
* Escape hatch
* Login query parameters
* Logout query parameters
* Enable PKCE